test: rename file
